+++ CHANGELOG.md
@@ -3,7 +3,7 @@ DAuth - ChangeLog
 
 (Dates below are YYYY/MM/DD)
 
-v0.6.1 - TBD - crypt(3)
+v0.6.1 - 2014/08/30 - Docs and crypt(3) formatting
 -------------------
 - **Enhancement:** Supports DMD 2.066.0. (Now supports DMD 2.064.2 through 2.066.0.)
 - **Enhancement:** Supports [crypt(3)](https://en.wikipedia.org/wiki/Crypt_(C))-style hash strings. Currently supports ```$1$``` (MD5), ```$5$``` (SHA-256) and ```$6$``` (SHA-512).